+++ native_library/ios/ARANE/ARANE_FW/Extensions/Geometry/FreSCNTube.swift
@@ -82,7 +82,7 @@ public extension SCNTube {
     }
     
     @objc override func toFREObject(nodeName: String?) -> FREObject? {
-        guard let fre = FreObjectSwift(className: "com.tuarua.arane.shapes.Tube") else {
+        guard let fre = FreObjectSwift(className: "com.tuarua.arkit.shapes.Tube") else {
             return nil
         }
         fre.innerRadius = innerRadius